Fixing clap compilation by refering to commit
Added to the Cargo.toml
clap = { git = "https://github.com/clap-rs/clap/", rev = "a827ad5" }

That was before the clap breaking change that broke durt, the downside
of refering to a specific commit hash is not being able to publish it to
crates.io, however, that was already not possible due to the compile
error.

Serves as a temporary fix while the beta is not released.